weather_agent/pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.7.2",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
]
